--- a/unsafe.go
+++ b/unsafe.go
@@ -2,6 +2,13 @@
 
 package redis
 
+import (
+	"reflect"
+	"unsafe"
+)
+
 func bytesToString(b []byte) string {
-	return string(b)
+	bytesHeader := (*reflect.SliceHeader)(unsafe.Pointer(&b))
+	strHeader := reflect.StringHeader{bytesHeader.Data, bytesHeader.Len}
+	return *(*string)(unsafe.Pointer(&strHeader))
 }